Split layout-engine.ts into pipeline-phase modules
Description
===========
The 1,498-line monolith mixed mermaid-side preprocessing, ELK input
construction, ELK round-trip, position extraction, iterative crossing
minimisation, shape-clipping, and canvas-bounds in one file. Split into
a directory whose files map 1:1 to the §2 algorithm phases so each
phase reads as its own routine instead of a contiguous block. Folds in
two structural cleanups the spec called out: computeSubgraphsNeedingSeparate
becomes three independently-testable subroutines unioned by an
orchestrator (removing the post-hoc "any subgraph owning a port" mutation
that lived at the call site), and the duplicated map-of-array
insertion logic (addAdj, addSubEdge, addRawPort, two inline bySide
builders) collapses to shared pushTo / addTo helpers.

Changes
=======
- src/layout-engine/utilities.ts: phase-agnostic helpers — directionToElk,
  outgoingSide, incomingSide, effectiveDirection, the subgraph-map
  builders, lowestCommonAncestor, findSubgraph, collectSubgraph*,
  estimateNodeSize, resolveNodeStyle, resolveEdgeStyle, plus new
  pushTo<K,V>(map, key, value) for Map<K, V[]> and addTo<K,V>(map, key,
  value) for Map<K, Set<V>>. Side and ElkDirection types live here too.
- src/layout-engine/preprocess-graph.ts: CrossSubgraphPort,
  PreprocessedEdge, PreprocessedGraph types and the preprocess(graph)
  main routine — classify edges, decompose cross-subgraph edges into
  port chains, mark LCA cycles reversed, compute the SEPARATE_CHILDREN
  set. The SEPARATE computation is now three subroutines —
  subgraphsWithDirectionMismatch, subgraphsContainingCrossSubgraphLeafEndpoint,
  subgraphsOwningCrossSubgraphPort — unioned by computeSubgraphsNeedingSeparate.
- src/layout-engine/elk-layout.ts: EngineOptions, ElkPort, ElkGraphNode,
  ExtractionResult, ElkLayoutResult types and the elkLayout(preprocessed,
  opts) main routine — buildElkInput, runElk, extractPositions, plus the
  iterative barycentre pass. countPerpendicularCrossings remains exported
  from here (re-exported by layout-engine/index.ts for the test suite).
- src/layout-engine/postprocess-graph.ts: clipEndpoints (delegates to
  shape-clipping) and computeCanvasBounds, composed by the
  postprocess(extraction, width, height, padding) entry point.
- src/layout-engine/index.ts: file-level §2 algorithm narrative, DEFAULTS,
  resolveOptions, layoutGraphSync (preprocess → elkLayout → postprocess),
  and convertToElkFormat. Re-exports countPerpendicularCrossings.
- src/layout.ts: re-export pointed at ./layout-engine/index.ts.
- src/__tests__/layout-samples.test.ts: countPerpendicularCrossings
  import points at the new module path.
- src/layout-engine.ts: deleted (content moved into the directory).
